ddhx: Add selection capability, support skip-* commands
Add the capacity to select multiple bytes.

This will enable commands (and shortcuts by extension) to perform actions on a selected range.

Selection information is shown in the status bar when active. `report-position` shows this on a byte/percentage basis.

For example, you can select a search criteria using shift+right-arrow, then skip this range (as a pattern) using ctrl+right-arrow. If [0x01,0x02,0x03,0x04] is selected, all aligned instances are checked and skipped until a different needle is met.